Extract mock DB classes
